src/main/java/nextstep/subway/line/domain/Sections.java
package nextstep.subway.line.domain;

import lombok.AccessLevel;
import lombok.NoArgsConstructor;
import nextstep.subway.station.domain.Station;

import javax.persistence.CascadeType;
import javax.persistence.Embeddable;
import javax.persistence.OneToMany;
import java.util.ArrayList;
import java.util.List;

@NoArgsConstructor(access = AccessLevel.PROTECTED)
@Embeddable
public class Sections {

    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
    private List<Section> sections = new ArrayList<>();

    public Sections(final List<Section> sections) {
        this.sections.addAll(sections);
    }

    public List<Station> getOrderedStations() {
        return new ArrayList<>();
    }

    public void add(final Section section) {
    }

    public void remove(final Station station) {

    }
}
